把local 中的集成测试放到单元测试中,redis 添加日志打印
