+++ middlewares/activelimit/redislimit/builder.go
@@ -1,6 +1,7 @@
 package redislimit
 
 import (
+	"fmt"
 	"github.com/gin-gonic/gin"
 	"github.com/redis/go-redis/v9"
 	"go.uber.org/atomic"
@@ -12,8 +13,9 @@ type RedisActiveLimit struct {
 	MaxActive *atomic.Int64
 
 	//用来记录连接数的key
-	key string
-	cmd redis.Cmdable
+	key   string
+	cmd   redis.Cmdable
+	logFn func(msg any, args ...any)
 }
 
 // NewRedisActiveLimit 全局限流
@@ -22,6 +24,9 @@ func NewRedisActiveLimit(cmd redis.Cmdable, maxActive int64, key string) *RedisA
 		MaxActive: atomic.NewInt64(maxActive),
 		key:       key,
 		cmd:       cmd,
+		logFn: func(msg any, args ...any) {
+			fmt.Println(fmt.Sprintf("%v  详细信息: %v", msg, args))
+		},
 	}
 }
 
@@ -30,20 +35,29 @@ func (a *RedisActiveLimit) SetMaxActive(maxActive int64) *RedisActiveLimit {
 	return a
 }
 
+func (a *RedisActiveLimit) SetLogFunc(fun func(msg any, args ...any)) *RedisActiveLimit {
+	a.logFn = fun
+	return a
+}
+
 func (a *RedisActiveLimit) Build() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 		currentCount, err := a.cmd.Incr(ctx, a.key).Result()
 		if err != nil {
 			//为了安全性 直接返回异常
+			a.logFn("redis 加一操作", err)
 			ctx.AbortWithStatus(http.StatusInternalServerError)
 			return
 		}
 		defer func() {
-			_ = a.cmd.Decr(ctx, a.key).Err()
+			if err = a.cmd.Decr(ctx, a.key).Err(); err != nil {
+				a.logFn("redis 减一操作", err)
+			}
 		}()
 		if currentCount <= a.MaxActive.Load() {
 			ctx.Next()
 		} else {
+			a.logFn("web server ", "限流中..")
 			ctx.AbortWithStatus(http.StatusTooManyRequests)
 		}
 		return
